requirements.txt
dotenv
streamlit
pillow
google.generativeai
datetime